새 폴더/10-01-django-rest-framework/assignment/django_ws_10_c/todos/serializers.py
# 재구성할 수 있는 포맷으로 변환하는 과정
from rest_framework import serializers
from .models import Todo

class TodoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Todo
        fields = '__all__'